football_data_analytics/requirements.txt
apache-airflow-providers-google==10.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
geographiclib==2.0
geopy==2.4.1
idna==3.6
nominatim==0.1
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2023.4
urllib3==2.1.0
